+++ common/src/main/java/net/spell_power/api/SpellSchools.java
@@ -39,10 +39,10 @@ public class SpellSchools {
         return new LinkedHashSet<SpellSchool>(REGISTRY.values());
     }
 
-
     // Predefined Spell Schools
 
-    public static final SpellSchool GENERIC = register(createMagic("generic", 0x9999BB, 100));
+    public static final SpellSchool GENERIC = register(createMagic("generic", 0x9999BB, 100))
+            .innateModifier(-1F);
     public static final SpellSchool ARCANE = register(createMagic("arcane", 0xff66ff));
     public static final SpellSchool FIRE = register(createMagic("fire", 0xff3300));
     public static final SpellSchool FROST = register(createMagic("frost", 0xccffff));