+++ common/src/main/java/net/spell_power/config/AttributesConfig.java
@@ -9,6 +9,7 @@ public class AttributesConfig {
         LIVING_ENTITY, PLAYER_ENTITY
     }
     // public AttributeScope attributes_container_injection_scope = AttributeScope.LIVING_ENTITY;
+    public float base_spell_power = 1F;
     public double base_spell_critical_chance_percentage = 5;
     public double base_spell_critical_damage_percentage = 50;
     public EffectConfig spell_power_effect = new EffectConfig(0.1F);